complete delete function of car

+++ Code/Client/admin/src/components/sections/transport-managment/cars/ViewCars.js
@@ -1,4 +1,5 @@
 import React, { useEffect, useState } from 'react';
+import Swal from 'sweetalert2';
 import axios from 'axios';
 import styles from "./viewcars.module.css";
 import { FaTrash, FaEdit } from "react-icons/fa";
@@ -21,23 +22,42 @@ export default function ViewCars() {
 
   // DELETE CAR DATA
 const deleteCar = (id) => {
-  axios.delete(`http://localhost:5000/api/cars/deleteCar/${id}`)
-    .then(res => {
-      console.log(res.data);
-      if (res.status === 200) {
-        axios.get("http://localhost:5000/api/cars/getCars")
-          .then(res => {
-            console.log(res.data);
-            setProduct(res.data.data);
-          })
-          .catch(err => {
-            console.log(err);
-          });
-      }
-    })
-    .catch(err => {
-      console.log(err);
-    });
+  Swal.fire({
+    title: 'Are you sure?',
+    text: "You want to delete this car!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, Delete!'
+  })
+
+  .then((result) => {
+    if (result.isConfirmed) {
+      axios.delete(`http://localhost:5000/api/cars/deleteCar/${id}`)
+      .then(res => {
+        Swal.fire(
+          'Delete Car!',
+          'You data have been delete.',
+          'success'
+        );
+        if (res.status === 200) {
+          axios.get("http://localhost:5000/api/cars/getCars")
+            .then(res => {
+              console.log(res.data);
+              setProduct(res.data.data);
+            })
+            .catch(err => {
+              console.log(err);
+            });
+        }
+      })
+      .catch(err => {
+        console.log(err);
+      });
+    }
+  });
+    
 }
